alert added

+++ src/Views/studentsHome.php
@@ -73,12 +73,13 @@ require_once("components/layoutHome.php")
                                     </div>
                                     </div>         
 
-                                    <a href='?action=delete&id={$student->getId()}'><img class='delete' src='public/img/iconDelete.svg' alt='delete'></a>
+                                    <a href='?action=delete&id={$student->getId()}'><img class='delete' src='public/img/iconDelete.svg' alt='delete' onclick='return confirmDelete();'></a>
                                 </td>
                             </tr>
                         ";
                     }
                     ?>
+                
                 </tbody>
             </table>
         </div>
@@ -94,6 +95,23 @@ require_once("components/layoutHome.php")
     ?>
 
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0/dist/js/bootstrap.bundle.min.js" integrity="sha384-A3rJD856KowSb7dwlZdYEkO39Gagi7vIsF0jrRAoQmDKKtQBHUuLZ9AsSv4jD4Xa" crossorigin="anonymous"></script>
+    
+
+    <script type="text/javascript">
+    function confirmDelete()
+    {
+        var respuesta = confirm("Are you sure you want to delete?");
+
+        if (respuesta == true) 
+        {
+            return true;
+        }
+        else
+        {
+            return false;
+        }
+    }
+</script>    
 </body>
 
 </html>
@@ -147,7 +165,7 @@ require_once("components/layoutHome.php")
         background-color: white;
         background-repeat: no-repeat;
         background-size: 10px 100%;
-        +
+        
     }
 
     #scrollYTable {
